requirements.txt
ultralytics
inference
pillow
tqdm
